+++ Cargo.toml
@@ -177,6 +177,7 @@ num-traits = "0.2"
 once_cell = "1.19.0"
 path_macro = "1.0.0"
 pretty_assertions = "1.4.0"
+proptest = "1.5.0"
 rand = "^0.8"
 rayon = "1.10.0"
 regex = "1.10.4"
